Documentaion/RESTapi/sample files/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
opencv-python
faster-whisper
langchain
langchain-openai
python-dotenv